Created RecommendedBooksWidget to display books in the home page.
