+++ lib/ui/pages/home/tab_home.dart
@@ -1,11 +1,11 @@
 import 'package:bookclub/domain/models/book.dart';
 import 'package:bookclub/domain/models/user.dart';
 import 'package:bookclub/generated/l10n.dart';
+import 'package:bookclub/ui/pages/home/widgets/recommended_books_widget.dart';
 import 'package:bookclub/ui/state/books_bloc/books_bloc.dart';
 import 'package:bookclub/ui/widgets/ui_avatar_card.dart';
 import 'package:bookclub/ui/widgets/ui_book_card.dart';
 import 'package:bookclub/ui/widgets/ui_page_header.dart';
-import 'package:bookclub/ui/widgets/ui_resizable_widget.dart';
 import 'package:flutter/material.dart' as material;
 import 'package:flutter/widgets.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -72,40 +72,13 @@ class TabHome extends StatelessWidget {
                         onPress: () {},
                       );
                     })),
-            UIPageHeader(title: S.of(context).recommended),
-            UIResizableWidget(onSize: (screenWidth, _) {
-              return _calculateGridSize(screenWidth);
-            }, onBuild: (context, size) {
-              int rowsQty = size.toInt();
-              return GridView.builder(
-                  gridDelegate: material.SliverGridDelegateWithFixedCrossAxisCount(
-                      crossAxisCount: rowsQty, crossAxisSpacing: 1.0, mainAxisSpacing: 1.0),
-                  primary: false,
-                  shrinkWrap: true,
-                  itemCount: recommendedBooks.length,
-                  itemBuilder: (ctx, index) {
-                    Book book = recommendedBooks[index];
-                    return UIBookCard(
-                      title: book.title,
-                      description: book.description ?? '',
-                      imageUrl: book.thumbnail ?? '',
-                      onPress: () {},
-                    );
-                  });
-            }),
+            RecommendedBooksWidget(
+              isLoading: false,
+              recommendedBooks: recommendedBooks,
+            ),
           ],
         ),
       ),
     );
-  }
-
-  double _calculateGridSize(double screenWidth) {
-    if (screenWidth > 600) {
-      return 4;
-    } else if (screenWidth > 400) {
-      return 3;
-    } else {
-      return 2;
-    }
   }
 }
